# fixtures/openapi.yaml
openapi: 3.1.0
info:
  title: Integration Test Fixture Contract
  version: 1.1.0
  description: |
    Language-agnostic functional contract for the integration-test fixture
    app. Implementations of this spec are generated as stubs for each
    supported language; every implementation must behave identically so
    the integration tests can validate the Wasmer virtualization backend
    across all deployment targets — the local platform, Edge, and Edge in
    the cloud — independently of the language the fixture is written in.

    Capabilities covered: instance liveness, environment introspection for
    managed databases, database connectivity, durable state via persistent
    counters, outbound HTTP (blocking and non-blocking), an aggregate
    self-test built for uptime probes, and a catch-all path echo with
    request logging.

    The `/ws` path is reserved for the WebSocket contract in
    `asyncapi.yaml` and is excluded from the catch-all. A fixture that
    implements that document must implement it at the same `info.version`;
    the two specs are one contract split by protocol.
servers:
  - url: "{appUrl}"
    description: >-
      Per-deployment base URL assigned when the fixture app is deployed to
      the target under test (local platform, Edge, or cloud Edge).
    variables:
      appUrl:
        default: http://localhost
security: []
...
paths:
...
  /async:
    post:
      tags: [outbound-http]
      operationId: proxyRequestNonBlocking
      summary: Perform an outbound HTTP request using non-blocking I/O
      description: >-
        Issues the requested outbound HTTP call using the implementation
        language's non-blocking / asynchronous I/O mechanism.
      requestBody:
        $ref: "#/components/requestBodies/ProxyRequest"
      responses:
        "200":
          $ref: "#/components/responses/ProxyResult"
  /sync:
    post:
      tags: [outbound-http]
      operationId: proxyRequestBlocking
      summary: Perform an outbound HTTP request using blocking I/O
      description: >-
        Issues the requested outbound HTTP call using the implementation
        language's blocking / synchronous I/O mechanism.
      requestBody:
        $ref: "#/components/requestBodies/ProxyRequest"
      responses:
        "200":
          $ref: "#/components/responses/ProxyResult"
  /self-test:
    get:
      tags: [self-test]
      operationId: selfTest
      summary: Run every inside-runnable contract check and report the outcome
      description: >-
        Executes the contract's capabilities from inside the instance and
        aggregates the outcome into one machine-readable report. Responds
        `200` only when every check passes and `500` otherwise, so a plain
        status-code validator can drive alerting; the body carries the same
        report either way.

        Check names are fixed across implementations: `db-env` (the
        getDbEnvReport data is internally consistent — the five required
        vars all present or all absent, never partially injected),
        `db-connect` (with credentials injected a real connection must
        succeed, mirroring checkDbConnection; without them the check passes
        as cleanly database-less), `counter-default` and `counter-named`
        (a read and an increment behave per durable-state; the named check
        uses the reserved counter name `self-test`), `counter-invalid-name`
        (a name outside `^[a-z-]+$` is rejected), and `echo` (the
        catch-all payload is well-formed, unique hash included).
        Implementations may not add further checks; a report is comparable
        across languages.

        Every check is verifiable purely from inside the instance; the
        self-test never opens a connection back to itself. Guest loopback
        is not routable on Edge, and any other outbound target would tie
        the probe's health to something outside the node — so the
        outbound-http and WebSocket capabilities are deliberately excluded
        here and covered by their own endpoints instead.

        The counter checks mutate durable state — each run increments the
        default counter and the `self-test` counter — every other check is
        side-effect free. Like `/ws`, this path is excluded from the
        catch-all and emits no catch-all log line.
      responses:
        "200":
          description: Every check passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelfTestReport"
        "500":
          description: At least one check failed; same report shape
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelfTestReport"
  /ws:
    get:
      tags: [websocket-upgrade]
      operationId: websocketUpgrade
      summary: WebSocket endpoint; not part of the catch-all
      description: >-
        Reserved by `asyncapi.yaml`, which defines everything that happens
        after a successful RFC 6455 handshake. Documented here only to fix
        the two behaviours the HTTP contract owns: a request without a
        valid upgrade is refused with `426 Upgrade Required`, and no
        request to this path — upgraded or not — emits the catch-all log
        line, so log assertions stay free of WebSocket traffic.
      responses:
        "101":
          description: >-
            Handshake accepted; the connection continues under
            `asyncapi.yaml`.
        "426":
          description: Request carried no valid WebSocket upgrade
          content:
            text/plain:
              schema:
                type: string
...
components:
...
  requestBodies:
    ProxyRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              method:
                type: string
                description: HTTP method for the outbound request
              target:
                type: string
                format: uri
                description: URL to request
              timeout_ms:
                type: number
                description: Request timeout in milliseconds
            required: [method, target, timeout_ms]
  responses:
...
    ProxyResult:
      description: >-
        Outcome of the outbound request. Errors (including timeouts) are
        reported in-band with HTTP 200 and an `error` field, so tests can
        always read the elapsed time.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  body:
                    type: string
                    description: Response body of the outbound request
                  status_code:
                    type: integer
                    description: Status code of the outbound request
                  elapsed_time_ms:
                    type: integer
                    description: Wall-clock duration of the outbound request
                required: [body, status_code, elapsed_time_ms]
              - type: object
                properties:
                  error:
                    type: string
                    description: Human-readable failure reason
                  status_code:
                    type: integer
                    const: 500
                  elapsed_time_ms:
                    type: integer
                    description: Wall-clock duration until the failure
                required: [error, status_code, elapsed_time_ms]
  schemas:
    SelfTestCheck:
      type: object
      properties:
        name:
          type: string
          description: Stable check identifier, unique within a report
        ok:
          type: boolean
        elapsed_ms:
          type: integer
          description: Wall-clock duration of the check
        error:
          type: string
          description: Failure reason; present only when `ok` is false
      required: [name, ok, elapsed_ms]
    SelfTestReport:
      type: object
      properties:
        ok:
          type: boolean
          description: True when every check passed
        checks:
          type: array
          items:
            $ref: "#/components/schemas/SelfTestCheck"
        unique_hash:
          type: string
          description: Per-deployment unique value injected at deploy time
      required: [ok, checks, unique_hash]
...